+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add `rreplace_format` for tabulation post-processing.
 * Add new tern function `t_ancova` to create ANCOVA tables, as well as corresponding elementary table function 
   `t_el_ancova` and summary function `s_ancova`.
+* Add new tern function `s_odds_ratio` to estimate Odds Ratio of response between categories.
 
 # tern 0.6.7